requirements.txt
django==5.1.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework==3.15.2
pillow==11.0.0
djangorestframework-simplejwt==5.3.1
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.0
coverage==7.6.8
requests==2.32.3
drf-yasg==1.21.8
pytz==2024.2
django-cors-headers==4.6.0
eventlet==0.38.0